DPROD-814: add 16 non-deprecated devfile/ repos for import
From talking with the team, they have some lower touch repos in devfile/ that we should include in Logilica for completeness. This adds them, and I'll need to run tool and do an import if there are no concerns. There are 21 repos in their org, 3 are marked deprecated and don't need to be imported, and 2 are already in our config. This PR adds the additional 16 repos.

Maps to DPROD-814